Converted readable streams
But tests are failing all over the place so there are probably some dumb mistakes in the conversion still
