+++ reference-implementation/lib/ReadableStreamBYOBRequest-impl.js
@@ -1,0 +1,30 @@
+'use strict';
+
+const { IsDetachedBuffer } = require('./abstract-ops/ecmascript.js');
+const aos = require('./abstract-ops/readable-streams.js');
+
+exports.implementation = class ReadableStreamBYOBRequestImpl {
+  get view() {
+    return this._view;
+  }
+
+  respond(bytesWritten) {
+    if (this._controller === undefined) {
+      throw new TypeError('This BYOB request has been invalidated');
+    }
+
+    if (IsDetachedBuffer(this._view.buffer) === true) {
+      throw new TypeError('The BYOB request\'s buffer has been detached and so cannot be used as a response');
+    }
+
+    aos.ReadableByteStreamControllerRespond(this._controller, bytesWritten);
+  }
+
+  respondWithNewView(view) {
+    if (this._controller === undefined) {
+      throw new TypeError('This BYOB request has been invalidated');
+    }
+
+    aos.ReadableByteStreamControllerRespondWithNewView(this._controller, view);
+  }
+};
