+++ reference-implementation/lib/ReadableStreamDefaultController.webidl
@@ -1,0 +1,8 @@
+[Exposed=(Window,Worker,Worklet)]
+interface ReadableStreamDefaultController {
+  readonly attribute unrestricted double? desiredSize;
+
+  void close();
+  void enqueue(optional any chunk);
+  void error(optional any e);
+};
