+++ reference-implementation/lib/UnderlyingSink.webidl
@@ -6,7 +6,7 @@ dictionary UnderlyingSink {
   any type;
 };
 
-callback WritableStreamStartCallback = Promise<void> (WritableStreamDefaultController controller);
+callback WritableStreamStartCallback = any (WritableStreamDefaultController controller);
 callback WritableStreamWriteCallback = Promise<void> (WritableStreamDefaultController controller, optional any chunk);
 callback WritableStreamCloseCallback = Promise<void> ();
 callback WritableStreamAbortCallback = Promise<void> (optional any reason);
