+++ reference-implementation/lib/abstract-ops/ecmascript.js
@@ -3,15 +3,13 @@ const assert = require('assert');
 
 const isFakeDetached = Symbol('is "detached" for our purposes');
 
-exports.typeIsObject = x => (typeof x === 'object' && x !== null) || typeof x === 'function';
-
-exports.createArrayFromList = elements => {
+exports.CreateArrayFromList = elements => {
   // We use arrays to represent lists, so this is basically a no-op.
   // Do a slice though just in case we happen to depend on the unique-ness.
   return elements.slice();
 };
 
-exports.ArrayBufferCopy = (dest, destOffset, src, srcOffset, n) => {
+exports.CopyDataBlockBytes = (dest, destOffset, src, srcOffset, n) => {
   new Uint8Array(dest).set(new Uint8Array(src, srcOffset, n), destOffset);
 };
 
@@ -35,15 +33,4 @@ exports.TransferArrayBuffer = O => {
 // Not implemented correctly
 exports.IsDetachedBuffer = O => {
   return isFakeDetached in O;
-};
-
-exports.rethrowAssertionErrorRejection = e => {
-  // Used throughout the reference implementation, as `.catch(rethrowAssertionErrorRejection)`, to ensure any errors
-  // get shown. There are places in the spec where we do promise transformations and purposefully ignore or don't
-  // expect any errors, but assertion errors are always problematic.
-  if (e && e instanceof assert.AssertionError) {
-    setTimeout(() => {
-      throw e;
-    }, 0);
-  }
 };
