+++ reference-implementation/lib/abstract-ops/internal-methods.js
@@ -2,3 +2,5 @@
 
 exports.AbortSteps = Symbol('[[AbortSteps]]');
 exports.ErrorSteps = Symbol('[[ErrorSteps]]');
+exports.CancelSteps = Symbol('[[CancelSteps]]');
+exports.PullSteps = Symbol('[[PullSteps]]');
